# docs/swagger.yaml
definitions:
...
  forms.PwdForm:
    properties:
      confirm_new_pwd:
        type: string
      new_pwd:
        type: string
      old_pwd:
        type: string
      username:
        type: string
    required:
    - confirm_new_pwd
    - new_pwd
    - old_pwd
    - username
    type: object
...
  forms.TagAddForm:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  forms.TagEditForm:
    properties:
      ID:
        type: integer
      name:
        type: string
    required:
    - ID
    - name
    type: object
...
  forms.UserInfoForm:
    properties:
      email:
        type: string
      nickname:
        type: string
      signature:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  utils.Result:
    properties:
      code:
        description: 状态码
        example: 100
        type: integer
      data:
        description: 数据
        type: object
      msg:
        description: 提示
        example: 密码错误
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  description: Aries 开源博客项目 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Swagger
  version: "1.0"
paths:
...
  /api/v1/all_tags:
    get:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取所有标签
      tags:
      - 标签
  /api/v1/all_users:
    get:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取所有用户
      tags:
      - 用户
...
  /api/v1/tags:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ids
        in: query
        name: ids
        required: true
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 批量删除标签
      tags:
      - 标签
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: size
        type: integer
      - description: 关键词
        in: query
        name: key
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 分页获取标签
      tags:
      - 标签
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加标签表单
        in: body
        name: addForm
        required: true
        schema:
          $ref: '#/definitions/forms.TagAddForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 添加标签
      tags:
      - 标签
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改标签表单
        in: body
        name: editForm
        required: true
        schema:
          $ref: '#/definitions/forms.TagEditForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 修改标签
      tags:
      - 标签
  /api/v1/tags/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 删除标签
      tags:
      - 标签
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 根据 ID 获取标签
      tags:
      - 标签
  /api/v1/users:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户信息表单
        in: body
        name: userForm
        required: true
        schema:
          $ref: '#/definitions/forms.UserInfoForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 更新用户信息
      tags:
      - 用户
  /api/v1/users/pwd:
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改密码表单
        in: body
        name: pwdForm
        required: true
        schema:
          $ref: '#/definitions/forms.PwdForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 修改密码
      tags:
      - 用户
swagger: "2.0"